% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Extract a contrast from a SELEX multinomial regression}
\usage{
contrast(fit, ref_allele, alt_allele)
}
\arguments{
\item{fit}{The multinom fit object from selex_multinom}

\item{ref_allele}{Character string indicating the reference allele}

\item{alt_allele}{Character string indicating the alternate allele}
}
\value{
Intercept and coefficient for the contrast
}
\description{
The contrast for a provided ref/alt allele pair
}
