% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_weights}
\alias{default_weights}
\title{The default weighting scheme for SELEX multinomial regression}
\format{An integer vector with 5 elements}
\usage{
default_weights
}
\description{
The cycle weights are 8, 1, 2, 4, 8
}
\keyword{datasets}
