% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_counts}
\alias{example_counts}
\title{An example set of count data for a SELEX experiment}
\format{A data frame with 5 rows and 4 columns}
\usage{
example_counts
}
\description{
An example set of count data for a SELEX experiment
}
\keyword{datasets}
