% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_regression.R
\name{preprocess_counts}
\alias{preprocess_counts}
\title{Preprocess count data}
\usage{
preprocess_counts(counts, ref_allele = NULL)
}
\arguments{
\item{counts}{A matrix or data frame with alleles for column names}

\item{ref_allele}{A character string representing the reference allele}
}
\value{
Matrix containing SELEX count data
}
\description{
Prepares count data for multinomial logistic regression
}
\details{
Converts the input data to matrix format if necessary and ensures that 
reference allele counts occupy the first column.
}
\seealso{
\code{\link{selex_multinom}}
}
