% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{randomize_counts}
\alias{randomize_counts}
\title{Randomize SELEX count data}
\usage{
randomize_counts(counts)
}
\arguments{
\item{counts}{A matrix or data frame with alleles for column names}
}
\value{
A matrix of randomized counts
}
\description{
Produces a randomized version of read counts from a SELEXexperiment
}
\seealso{
\code{\link{sample_coefficients}}
}
