% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{sample_coefficients}
\alias{sample_coefficients}
\title{Generate a random sample of SELEX multinomial regression coefficients}
\usage{
sample_coefficients(counts, weights = default_weights, n = 1,
  cores = detectCores())
}
\arguments{
\item{counts}{A matrix or data frame with alleles for column names}

\item{weights}{A numeric vector of regression weights for the 5 cycles}

\item{n}{An integer giving number of coefficient samples to generate}

\item{cores}{An integer giving the number of cpu cores to use}
}
\value{
A matrix of coefficients as drawn from the null distribution
}
\description{
Randomizes the input counts and generates null coefficients
}
\seealso{
\code{\link{estimate_standard_errors}}
}
