% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{selex_pvals}
\alias{selex_pvals}
\title{Estimate p-values for SELEX multinomial regression coefficients}
\usage{
selex_pvals(fit, estimated_se = NULL, n = 100, cores = detectCores(),
  timeout = 1)
}
\arguments{
\item{fit}{A matrix or data frame with alleles for column names}

\item{estimated_se}{A numeric vector giving standard errors to use}

\item{n}{An integer giving number of coefficient samples to generate}

\item{cores}{An integer giving the number of cpu cores to use}

\item{timeout}{Numeric. The time limit for fitting a model during null
sampling, in seconds. (workaround for divergent edge cases)}
}
\value{
A numeric vector containing the estimated p-values
}
\description{
Compares multinomial logistic regression parameters to estimated SE's
}
