% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{selex_standard_errors}
\alias{selex_standard_errors}
\title{Estimate standard errors for multinomial regression coefficients}
\usage{
selex_standard_errors(counts, weights = default_weights, n = 100,
  cores = detectCores())
}
\arguments{
\item{counts}{A matrix or data frame with alleles for column names}

\item{weights}{A numeric vector of regression weights for the 5 cycles}

\item{n}{An integer giving number of coefficient samples to generate}

\item{cores}{An integer giving the number of cpu cores to use}
}
\value{
A numeric vector containing the estimated SE's
}
\description{
Generates a sample of randomized SELEX data and computes "null" coefficients
}
\seealso{
\code{\link{selex_z_scores}}
}
