% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{selex_z_scores}
\alias{selex_z_scores}
\title{Estimate Z-scores for SELEX multinomial regression coefficients}
\usage{
selex_z_scores(fit, estimated_se = NULL, n = 100, cores = detectCores())
}
\arguments{
\item{fit}{A matrix or data frame with alleles for column names}

\item{estimated_se}{A numeric vector giving standard errors to use}

\item{n}{An integer giving number of coefficient samples to generate}

\item{cores}{An integer giving the number of cpu cores to use}
}
\value{
A numeric vector containing the estimated Z-scores
}
\description{
Compares multinomial logistic regression parameters to estimated SE's
}
\seealso{
\code{\link{selex_pvals}}
}
