% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{two_tailed_z_test}
\alias{two_tailed_z_test}
\title{A two-tailed Z-test}
\usage{
two_tailed_z_test(z)
}
\arguments{
\item{z}{A Z-score}
}
\value{
A p-value
}
\description{
A two-tailed Z-test
}
\seealso{
\code{\link{selex_pvals}}
}
